move files from parser subpackage to molecule package
